wip: remove test code and add to mining button
